Disambiguation/disambiguationProgram/requirements.txt
bottle==0.12.25
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
proxy-tools==0.1.0
pycairo==1.23.0
PyGObject==3.44.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywebview==4.1
regex==2023.6.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
